requirements.txt
# apt-get update && apt-get install -y portaudio19-dev - Linux
# brew install portaudio - Mac
pydub
streamlit 
googletrans=4.0.0
google-cloud-speech
google-cloud-texttospeech
transformers 
sentencepiece
torch 
torchvision 
torchaudio